pyproject.toml
[tool.poetry]
name = "iot_project"
version = "0.1.0"
description = ""
authors = ["Kensei Sakai <kensei.sakai@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"
uvicorn = "^0.17.0"
fastapi = "^0.71.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
